Re-engineered the whole Unit thing. KWUnit isn't a class anymore but a bunch of utility functions. Values are stored in pt, as doubles (instead of in the 3 units!), and are rounded when being converted to the user's unit before being displayed.
svn path=/trunk/koffice/; revision=93298

+++ kword/paragdia_p.h
@@ -39,23 +39,23 @@ public:
     KWPagePreview( QWidget*, const char* = 0L );
     ~KWPagePreview() {}
 
-    void setLeft( float _left )
+    void setLeft( double _left )
     { left = _left; repaint( false ); }
-    void setRight( float _right )
+    void setRight( double _right )
     { right = _right; repaint( false ); }
-    void setFirst( float _first )
+    void setFirst( double _first )
     { first = _first; repaint( false ); }
-    void setSpacing( float _spacing )
+    void setSpacing( double _spacing )
     { spacing = _spacing; repaint( false ); }
-    void setBefore( float _before )
+    void setBefore( double _before )
     { before = _before; repaint( false ); }
-    void setAfter( float _after )
+    void setAfter( double _after )
     { after = _after; repaint( false ); }
 
 protected:
     void drawContents( QPainter* );
 
-    float left, right, first, spacing, before, after;
+    double left, right, first, spacing, before, after;
 
 };
